Update the path following algorithm to be more efficient.
The previous algorithm I used simply used the law of cosines and some hacky reasoning to determine where it needed to turn.
It ended up not hitting walls, but it just spun in circles until it got everywhere.
This new algorithm gets the cross product of vectors to determine the sign of the angle of rotation, proving much more consistent.
I've also made some other changes through the algorithm which seem to have helped it be more consistent.

There is still some tuning that needs to be done with PID.